+++ changelog.md
@@ -5,8 +5,9 @@
 ## changelog
 
 ### 20260323
-matmul gpu tile版本初步跑通，待进一步优化，以及和标准方式对比。
-flash attention cpu 版本reference就绪。
+1. matmul gpu tile版本初步跑通，待进一步优化，以及和标准方式对比。
+2. flash attention cpu 版本reference就绪。
+3. flash attention backend 框架ai生成，vllm可以识别，但不能跑。
 
 ### 20260322
 matmul gpu tile版本有逻辑错误，待确认正确方式。